+++ spec/requests/as_content_owner/video_spec.rb
@@ -190,6 +190,17 @@ describe Yt::Video, :partner do
         end
       end
 
+      describe 'views can be grouped by state' do
+        let(:range) { {since: 4.days.ago, until: 3.days.ago} }
+
+        specify 'with the :by option set to :state' do
+          views = video.views range.merge by: :state
+          expect(views.keys).to all(be_a String)
+          expect(views.keys.map(&:length).uniq).to eq [2]
+          expect(views.values).to all(be_an Integer)
+        end
+      end
+
       describe 'comments can be retrieved for a specific day' do
         context 'in which the video was partnered' do
           let(:comments) { video.comments_on ENV['YT_TEST_PARTNER_VIDEO_DATE']}
@@ -853,6 +864,17 @@ describe Yt::Video, :partner do
         end
       end
 
+      describe 'estimated minutes watched can be grouped by state' do
+        let(:range) { {since: 4.days.ago, until: 3.days.ago} }
+
+        specify 'with the :by option set to :state' do
+          minutes = video.estimated_minutes_watched range.merge by: :state
+          expect(minutes.keys).to all(be_a String)
+          expect(minutes.keys.map(&:length).uniq).to eq [2]
+          expect(minutes.values).to all(be_a Float)
+        end
+      end
+
       describe 'average view duration can be retrieved for a specific day' do
         context 'in which the video was partnered' do
           let(:average_view_duration) { video.average_view_duration_on 5.days.ago}
@@ -921,6 +943,17 @@ describe Yt::Video, :partner do
         end
       end
 
+      describe 'average view duration can be grouped by state' do
+        let(:range) { {since: 4.days.ago, until: 3.days.ago} }
+
+        specify 'with the :by option set to :state' do
+          duration = video.average_view_duration range.merge by: :state
+          expect(duration.keys).to all(be_a String)
+          expect(duration.keys.map(&:length).uniq).to eq [2]
+          expect(duration.values).to all(be_a Float)
+        end
+      end
+
       describe 'average view percentage can be retrieved for a specific day' do
         context 'in which the video was partnered' do
           let(:average_view_percentage) { video.average_view_percentage_on 5.days.ago}
@@ -983,6 +1016,17 @@ describe Yt::Video, :partner do
 
         specify 'with the :by option set to :country' do
           percentage = video.average_view_percentage range.merge by: :country
+          expect(percentage.keys).to all(be_a String)
+          expect(percentage.keys.map(&:length).uniq).to eq [2]
+          expect(percentage.values).to all(be_a Float)
+        end
+      end
+
+      describe 'average view percentage can be grouped by state' do
+        let(:range) { {since: 4.days.ago, until: 3.days.ago} }
+
+        specify 'with the :by option set to :state' do
+          percentage = video.average_view_percentage range.merge by: :state
           expect(percentage.keys).to all(be_a String)
           expect(percentage.keys.map(&:length).uniq).to eq [2]
           expect(percentage.values).to all(be_a Float)
@@ -1173,6 +1217,17 @@ describe Yt::Video, :partner do
         end
       end
 
+      describe 'annotation clicks can be grouped by state' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :state' do
+          clicks = video.annotation_clicks range.merge by: :state
+          expect(clicks.keys).to all(be_a String)
+          expect(clicks.keys.map(&:length).uniq).to eq [2]
+          expect(clicks.values).to all(be_an Integer)
+        end
+      end
+
       describe 'annotation click-through rate can be retrieved for a range of days' do
         let(:date) { ENV['YT_TEST_PARTNER_VIDEO_DATE'] }
         let(:date_to) { Date.parse(ENV['YT_TEST_PARTNER_VIDEO_DATE']) + 5 }
@@ -1221,6 +1276,17 @@ describe Yt::Video, :partner do
         end
       end
 
+      describe 'annotation click-through rate can be grouped by state' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :state' do
+          rate = video.annotation_click_through_rate range.merge by: :state
+          expect(rate.keys).to all(be_a String)
+          expect(rate.keys.map(&:length).uniq).to eq [2]
+          expect(rate.values).to all(be_a Float)
+        end
+      end
+
       describe 'annotation close rate can be retrieved for a range of days' do
         let(:date) { ENV['YT_TEST_PARTNER_VIDEO_DATE'] }
         let(:date_to) { Date.parse(ENV['YT_TEST_PARTNER_VIDEO_DATE']) + 5 }
@@ -1263,6 +1329,17 @@ describe Yt::Video, :partner do
 
         specify 'with the :by option set to :country' do
           rate = video.annotation_close_rate range.merge by: :country
+          expect(rate.keys).to all(be_a String)
+          expect(rate.keys.map(&:length).uniq).to eq [2]
+          expect(rate.values).to all(be_a Float)
+        end
+      end
+
+      describe 'annotation close rate can be grouped by state' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :state' do
+          rate = video.annotation_close_rate range.merge by: :state
           expect(rate.keys).to all(be_a String)
           expect(rate.keys.map(&:length).uniq).to eq [2]
           expect(rate.values).to all(be_a Float)