+++ CHANGELOG.md
@@ -9,6 +9,7 @@ For more information about changelogs, check
 ## 0.22.1 - unreleased
 
 * [FEATURE] New `by: :country` option for channel, video and playlist reports
+* [FEATURE] New `by: :state` option for channel, video and playlist reports
 
 ## 0.22.0 - 2015-04-30